+++ changelog.md
@@ -1,3 +1,12 @@
+# PIC24 DSPIC33 FREERTOS DEMO v1.1.0
+### Release Highlights
+
+Demos have been updated to use the FreeRTOS Kernel version 10.5.1.
+
+
+### Features Added\Updated
+dsPIC33C Client demo has been updated to include a task to demonstrate the MSI feature.
+
 # PIC24 DSPIC33 FREERTOS DEMO v1.0.0
 ### Release Highlights